fix: Resolve latent bugs and clang errors found by clangd
- CompileShaders passed std::string through Logger::Log varargs; it only
  worked because MSVC stores the heap pointer first in long strings.
- The ortho shadow direction was a D3DXVECTOR4 built from a D3DXVECTOR3,
  reading w past the end of the temporary. Construct it explicitly.
- Logger takes const char*, so ternaries of string literals compile.
- Cast enums before cross-enum arithmetic (ill-formed in C++26) and name
  vector temporaries instead of taking their address.

With the code fixed, the compile database no longer needs clang
warning suppressions.

+++ TESReloaded/Core/SettingManager.h
@@ -53,7 +53,7 @@ static const char* IntroMovie = "";
 static const char* MainMenuMovie = "";
 static const char* MainMenuMusic = "";
 #endif
-#define WeatherColorsSize TESWeather::kNumColorTypes * TESWeather::kNumTimeOfDay * 4
+#define WeatherColorsSize ((UInt32)TESWeather::kNumColorTypes * TESWeather::kNumTimeOfDay * 4)
 #define SettingStringBuffer 80
 
 static const UInt32*	SettingGridsToLoad				= (UInt32*)kSettingGridsToLoad;
